requirements.txt
torch==1.10.1+cu102 torchvision==0.11.2+cu102 torchaudio===0.10.1+cu102 -f https://download.pytorch.org/whl/cu102/torch_stable.html
tensorboard
torchtext==0.6.0
openai
SpeechRecognition
Tk 
torchtext


